remove the 'Storage' impls for 'MaybeUninit'
They complicate the safety requirements necessary for the 'Storage'
traits, because the caller needed to guarantee that the 'self' provided
to the unsafe 'Storage' methods was "initialized enough" to be used;
now, we can define all values of a type implementing 'Storage' to be
"initialized enough".  This also means that ring buffers can now safely
accept any 'Storage' during construction.